Enum tutil::crayon::Color

pub enum Color {
    Black,
    Red,
    Green,
    Yellow,
    Blue,
    Purple,
    Cyan,
    White,
    Fixed(u8),
    Rgb(u8, u8, u8),
}

A Color is a specific ANSI colour name which can refer to either the foreground or background.

It can also be a custom value from 0 to 255 via the Fixed(u8) variant for terminals that have 256-colour. True-colour is also supported via the Rgb(u8, u8, u8) variant.

For further reading visit this Wikipedia page.

Variants

Black

Foreground code 30, background code 40.

Red

Foreground code 31, background code 41.

Green

Foreground code 32, background code 42.

Yellow

Foreground code 33, background code 43.

Blue

Foreground code 34, background code 44.

Purple

Foreground code 35, background code 45.

Cyan

Foreground code 36, background code 46.

White

Foreground code 37, background code 47.

Fixed

A value from 0 to 255 for use on terminals that have 256-colour support.

0 to 7 are the Black through White variants. These colours can usually be changed in the terminal emulator.

8 to 15 are brighter versions of the aforementioned colours. These colours can also usually be changed in the terminal emulator, however, it also could be configured to use the original colours and show the text in bold.

16 to 231 contain several palettes of bright colours, arranged in six squares measuring six by six each.

232 to 255 are shades of grey from black to white.

You can view this colour chart for a visual representation.

Rgb

A red value, green value and blue value, each between 0 to 255 (the RGB colour model) for use on terminals that have true-colour support.

Please note that true-colour support on many terminal emulators is patchy, however a fair share of the most common terminal emulators support it, such as:

Gnome Terminal

Konsole

Terminator

iTerm 2

For an up-to-date list of true-colour support visit: https://gist.github.com/XVilka/8346728.

Methods

impl Color

fn paint<'a, S>(self, string: S) -> StyledString<'a> where S: Into<Cow<'a, str>>

Convenience method for creating a StyledString with the foreground set without having to manually create a Style or use <color>.normal().paint().

fn normal(self) -> Style

Returns a Style with the foreground colour set to this colour.

fn on(self, background: Color) -> Style

The same as Color::normal(), but also sets the background colour.

fn bold(self) -> Style

Returns a Style with the 'bold' property set and the foreground colour set to this colour.

fn dimmed(self) -> Style

Returns a Style with the 'dimmed' property set and the foreground colour set to this colour.

fn italic(self) -> Style

Returns a Style with the 'italic' property set and the foreground colour set to this colour.

fn underline(self) -> Style

Returns a Style with the 'underline' property set and the foreground colour set to this colour.

fn blink(self) -> Style

Returns a Style with the 'blink' property set and the foreground colour set to this colour.

fn reverse(self) -> Style

Returns a Style with the 'reverse' property set and the foreground colour set to this colour.

fn hidden(self) -> Style

Returns a Style with the 'hidden' property set and the foreground colour set to this colour.