Struct tutil::crayon::Style

pub struct Style {
    // some fields omitted
}

A collection of properties that are used to format a string.

Methods

impl Style

fn new() -> Style

Creates a new Style without any formatting.

fn paint<'a, S>(self, string: S) -> StyledString<'a> where S: Into<Cow<'a, str>>

Applies the Style to a string, yielding a StyledString.

fn foreground(&self, color: Color) -> Style

Sets the foreground to the given colour.

fn background(&self, color: Color) -> Style

Sets the background to the given colour.

fn bold(&self) -> Style

Applies the 'bold' property.

fn dimmed(&self) -> Style

Applies the 'dimmed' property.

fn italic(&self) -> Style

Applies the 'italic' property.

fn underline(&self) -> Style

Applies the 'underline' property.

fn blink(&self) -> Style

Applies the 'blink' property.

fn reverse(&self) -> Style

Applies the 'reverse' property.

fn hidden(&self) -> Style

Applies the 'hidden' property.